refactor: apply barrel exports and update import paths (#18)
