--- src/components/index.ts
@@ -1,5 +1,0 @@
-// components 폴더의 공통 export를 한곳에서 관리
-
-export { default as Footer } from './layouts/Footer'
-export { default as Header } from './layouts/Header'
-export { default as RootLayout } from './layouts/RootLayout'
