+++ src/components/layouts/index.ts
@@ -1,0 +1,5 @@
+// components 폴더의 공통 export를 한곳에서 관리
+
+export { default as Footer } from './Footer'
+export { default as Header } from './Header'
+export { default as RootLayout } from './RootLayout'
